Fix spelling error in file name
